Commit refactored JS and HTML - View Album Modal
The JS code for populating a modal with album and user rating
information has been refactored.  I made improvements to the backend
Rails API and updated the API JS accordingly.  I also added html changes
to the modal layout to allow for alerts and better layout of the data
presented in the modal.